Added conveniences for encode/decode with string/byte[] types flipped.
This formulation is useful for doing any crypto work.
Basically you get a String input, need to encode it
to base64 so the encoding doesn't get mangled by the
cipher, use the resulting bytes for cipher and the
decode in reverse, bytes back to the original string.

Diff is a little bit of a mess as there was a bunch
of extra whitespace at the end of the lines that I
trimmed off (or editor did), and there is a mix of
tabs and spaces in the original file. I just went
with the flow and added tabs where they were before.